util: factor out CopyFile function
